refactor(http): restructure src/ into dispatch/, middleware/, adapters/, context/ subfolders
Move dispatch pipeline files to dispatch/, middleware files to middleware/,
binding/validation adapters to adapters/, and request-context/SSE to context/.
Update all internal import paths accordingly. No logic or type changes.

Ref: #850
