+++ packages/http/src/dispatch/dispatcher-status.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest';
 
 import { Container } from '@konekti/di';
 
-import type { FrameworkRequest, FrameworkResponse, HandlerDescriptor, HandlerMapping, Interceptor, InterceptorContext } from './types.js';
+import type { FrameworkRequest, FrameworkResponse, HandlerDescriptor, HandlerMapping, Interceptor, InterceptorContext } from '../types.js';
 import { createDispatcher } from './dispatcher.js';
 
 function createResponse(): FrameworkResponse & { body?: unknown } {
