+++ packages/http/src/middleware/cors.ts
@@ -1,4 +1,4 @@
-import type { Middleware } from './types.js';
+import type { Middleware } from '../types.js';
 
 export interface CorsOptions {
   allowCredentials?: boolean;
