+++ packages/http/src/middleware/middleware.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { Container } from '@konekti/di';
 
-import type { FrameworkResponse, Middleware, MiddlewareContext, MiddlewareLike, Next } from './types.js';
+import type { FrameworkResponse, Middleware, MiddlewareContext, MiddlewareLike, Next } from '../types.js';
 import { forRoutes, isMiddlewareRouteConfig, runMiddlewareChain } from './middleware.js';
 
 function createResponse(): FrameworkResponse {
